% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_pathways.R
\name{analyze_pathways}
\alias{analyze_pathways}
\title{Analyze Network Pathways}
\usage{
analyze_pathways(type_a, type_b, dir_out, depth, ntrial)
}
\arguments{
\item{type_a}{Cell type A}

\item{type_b}{Cell type B}

\item{dir_out}{Output directory}

\item{depth}{How many steps out to form neighborhood?}

\item{ntrial}{How many empirical simulations to run?}
}
\value{
None
}
\description{
Starting at each ligand-receptor pair in the final network, analyze small
neighbor graphs around these pairs. Consider them to be subsets of the larger
graph.
}
\examples{
{
type_a <- "BCells"
type_b <- "TCells"
dir_out <- "~/CytoTalk-output"
depth <- 3
ntrial <- 1000

analyze_pathways(type_a, type_b, dir_out, depth, ntrial)
}

}
