% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{check_count_data}
\alias{check_count_data}
\title{Transform Count Data if Detected}
\usage{
check_count_data(mat, auto_transform = TRUE)
}
\arguments{
\item{mat}{A numerical matrix}

\item{auto_transform}{Should the data be transformed if counts are detected?}
}
\description{
Transform Count Data if Detected
}
