% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_valid_names}
\alias{check_valid_names}
\title{Extract Valid Cell Types}
\usage{
check_valid_names(dir_in)
}
\arguments{
\item{dir_in}{Input directory, contains scRNAseq files}
}
\value{
A character vector, includes only valid scRNAseq cell types
}
\description{
Searches input directory for a particular regular expression pattern,
filtering out invalid filenames, and finally returning a list of valid cell
types.
}
\examples{
{
dir_in <- "scRNAseq-data"
check_valid_names(dir_in)
}

}
