% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_gene_network.R
\name{compute_gene_relevance}
\alias{compute_gene_relevance}
\title{Compute Gene Relevance}
\usage{
compute_gene_relevance(ligands, dir_in, dir_out)
}
\arguments{
\item{ligands}{Character matrix of ligand-receptor pair names}

\item{dir_in}{Input directory, contains scRNAseq files}

\item{dir_out}{Output directory}
}
\value{
None
}
\description{
Using a randomized walk with restart, compute gene relevance.
}
\examples{
{
ligands <- CytoTalk::ligands_mouse
dir_in <- "scRNAseq-data"
dir_out <- "CytoTalk-output"

compute_gene_relevance(ligands, dir_in, dir_out)
}

}
