% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mutual_information.R
\name{compute_mutual_information}
\alias{compute_mutual_information}
\title{Compute Mutual Information Matrix}
\usage{
compute_mutual_information(dir_out)
}
\arguments{
\item{dir_out}{Output directory}
}
\value{
None
}
\description{
Within each cell type, compute intracellular mutual information.
}
\examples{
\dontrun{
dir_out <- "my-output"
compute_mutual_information(dir_out)
}

}
