% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_non_self_talk.R
\name{compute_non_self_talk}
\alias{compute_non_self_talk}
\title{Compute Non-Self-Talk Score}
\usage{
compute_non_self_talk(ligands, type_a, type_b, dir_in, dir_out)
}
\arguments{
\item{ligands}{Character matrix of ligand-receptor pair names}

\item{type_a}{Cell type A}

\item{type_b}{Cell type B}

\item{dir_in}{Input directory, contains scRNAseq files}

\item{dir_out}{Output directory}
}
\value{
NIL
}
\description{
Given cell types and the names of ligand-receptor pairs, compute the
non-self-talk score (within each cell type).
}
