% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pem.R
\name{compute_pem}
\alias{compute_pem}
\title{Compute Preferential Expression Measure}
\usage{
compute_pem(dir_in, dir_out)
}
\arguments{
\item{dir_in}{Input directory, contains scRNAseq files}

\item{dir_out}{Output directory}
}
\value{
None
}
\description{
Using all cell types in the input folder, compute the PEM; this is done per
gene, per cell type.
}
\examples{
{
dir_in <- "~/scRNAseq-data"
dir_out <- "~/CytoTalk-output"

compute_pem(dir_in, dir_out)
}

}
