% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytotalk.R
\docType{package}
\name{cytotalk}
\alias{cytotalk}
\title{CytoTalk: Construct signal transduction networks}
\description{
The CytoTalk package allows for \emph{de novo} construction of a signaling network
(pathways from ligand-receptor pairs) between two cell types using single-
cell transcriptomics data (scRNA-seq).
}
\section{Running CytoTalk}{

The central wrapper function, \code{run_cytotalk}, is the main entry point to the
package. Most users should start here. Check out its MAN page by executing
\code{?run_cytotalk} in your R console.
}

\section{Advanced Users}{

If you're willing to do some digging, CytoTalk is composed of many steps,
some of which could be interchangable with other computational methods. For
example, let's say that you have a different idea for computing intra-
cellular similarity (i.e. not a mutual information matrix). You could skip
steps 1-3 and attempt to run from step 4 onward (simply execute
\code{run_cytotalk} in your console to view the source code) with differently
integrated data. Currently, doing so is not user-friendly, so this is
recommended for developers only. However, it would be interesting to make the
different components of the overall process more modular to compare
subprocesses.
}

