% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/fileio.R
\name{doc_fileio}
\alias{doc_fileio}
\alias{read_matrix_folder}
\alias{read_matrix_with_meta}
\title{File Input and Ouput}
\usage{
read_matrix_folder(
  dpath,
  pattern = ".*scRNAseq_(.+?)\\\\..+",
  auto_transform = TRUE
)

read_matrix_with_meta(fpath_mat, fpath_meta, auto_transform = TRUE)
}
\arguments{
\item{dpath}{The path of a directory, which contains scRNAseq matrices}

\item{pattern}{A regular expression, matches scRNAseq filenames}

\item{auto_transform}{Should count data be transformed if detected?}

\item{fpath_mat}{The path of a file containing a scRNAseq data matrix}

\item{fpath_meta}{The path of a file contianing column metadata (cell types)}
}
\description{
File Input and Ouput
}
