% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/graphing.R
\name{doc_graphing}
\alias{doc_graphing}
\alias{write_network_sif}
\alias{graph_pathway}
\title{Graphing with Cytoscape and Graphviz}
\usage{
write_network_sif(df_net, cell_type_a, dir_out)

graph_pathway(df_net_sub)
}
\arguments{
\item{df_net}{Final network; for example, the output of the
\code{extract_network} function}

\item{cell_type_a}{Name of cell type A that matches scRNA-seq file; for
example, \code{"Fibroblasts"}}

\item{dir_out}{Folder used for output}

\item{df_net_sub}{A subset of the final network (pathway); for example, the
output of the \code{extract_pathways} function}
}
\value{
Graphs which represent the subset pathways of the final network
}
\description{
Graphing with Cytoscape and Graphviz
}
\examples{
{
pathways <- CytoTalk::result_cyto$pathways$raw
result <- graph_pathway(pathways[[1]])
}

}
