% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/integrated.R
\name{doc_integrated}
\alias{doc_integrated}
\alias{nonselftalk}
\alias{gene_relevance}
\alias{node_prize}
\alias{crosstalk}
\alias{integrate_network}
\title{Integrated Co-Expression Network}
\usage{
nonselftalk(mat_type, lrp)

gene_relevance(mat_intra, lrp)

node_prize(mat_pem, cell_type, vec_relev)

crosstalk(
  mat_pem,
  cell_type_a,
  cell_type_b,
  vec_nst_a,
  vec_nst_b,
  mat_type,
  lrp
)

integrate_network(
  vec_nst_a,
  vec_nst_b,
  mat_intra_a,
  mat_intra_b,
  cell_type_a,
  cell_type_b,
  mat_pem,
  mat_type,
  lrp
)
}
\arguments{
\item{mat_type}{A matrix of a single cell type}

\item{lrp}{A dataframe or matrix object with two columns, ligands names and
the names of their receptors; by default, uses the \code{lrp_human} data. This
package also includes \code{lrp_mouse}, but you can also use your own data}

\item{mat_intra}{Intracellular network of any type}

\item{mat_pem}{PEM output}

\item{cell_type}{Name of any cell type}

\item{vec_relev}{Vector containing gene relevance scores}

\item{cell_type_a}{Name of cell type A that matches scRNA-seq file; for
example, \code{"Fibroblasts"}}

\item{cell_type_b}{Name of cell type B that matches scRNA-seq file; for
example, \code{"LuminalEpithelialCells"}}

\item{vec_nst_a}{Vector containing nonselftalk-scores for cell type A}

\item{vec_nst_b}{Vector containing nonselftalk-scores for cell type B}

\item{mat_intra_a}{Intracellular network for cell type A; for example, using
\code{parmigene::aracne.m} on the result of \code{mi_mat_parallel} will filter out
any indirect edges per the data processing inequality}

\item{mat_intra_b}{Intracellular network for cell type B}
}
\value{
Various outputs to build the integrated network
}
\description{
Integrated Co-Expression Network
}
\examples{
{
lst_scrna <- CytoTalk::scrna_cyto
cell_type_a <- "Macrophages"
lrp <- CytoTalk::lrp_human
mat_a <- extract_group(cell_type_a, lst_scrna)
result <- nonselftalk(mat_a, lrp)
}

}
