% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/pem.R
\name{doc_pem}
\alias{doc_pem}
\alias{pem_basic}
\alias{pem}
\title{Preferential Expression Measure}
\usage{
pem_basic(mat_scrna, labels)

pem(lst_scrna)
}
\arguments{
\item{mat_scrna}{Matrix containing scRNA-seq data, along with \code{labels},
defines the transformed count data and the associated cell types}

\item{labels}{Associated cell types, column-wise, of \code{mat_scrna}}

\item{lst_scrna}{List containing scRNA-seq data; for example, lists returned
from \code{read_matrix_folder} or \code{read_matrix_with_meta}}
}
\value{
Preferential expression measure scores
}
\description{
Preferential Expression Measure
}
\examples{
{
lst_scrna <- CytoTalk::scrna_cyto
result <- pem(lst_scrna)
}

}
