% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_gene_network.R
\name{extract_intracell_edges}
\alias{extract_intracell_edges}
\title{Extract Intracelluar Edges}
\usage{
extract_intracell_edges(dir_out)
}
\arguments{
\item{dir_out}{Output directory}
}
\description{
From the mutual information matrix, extract all non-zero edges and save them
in a data table.
}
