% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pathways.R
\name{generate_pathways}
\alias{generate_pathways}
\title{Generate Network Pathways}
\usage{
generate_pathways(type_a, type_b, dir_out, depth)
}
\arguments{
\item{type_a}{Cell type A}

\item{type_b}{Cell type B}

\item{dir_out}{Output directory}

\item{depth}{How many steps out to form neighborhood?}
}
\value{
None
}
\description{
Starting at each ligand-receptor pair in the final network, generate small
neighbor graphs around these pairs. Consider them to be subsets of the larger
graph.
}
\examples{
\dontrun{
type_a <- "BCells"
type_b <- "TCells"
dir_out <- "~/CytoTalk-output"
depth <- 3

generate_pathways(type_a, type_b, dir_out, depth)
}

}
