% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_signaling_network.R
\name{generate_signaling_network}
\alias{generate_signaling_network}
\title{Generate Signaling Network}
\usage{
generate_signaling_network(dir_out)
}
\arguments{
\item{dir_out}{Output directory}
}
\description{
From PCST runs, find the "best" choice for a network based on how unlikely
their distribution of edge counts is (higher the better).
}
