% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_gene_network.R
\name{integrate_network}
\alias{integrate_network}
\title{Integrate Cellular Network}
\usage{
integrate_network(ligands, type_a, type_b, dir_in, dir_out)
}
\arguments{
\item{ligands}{Character matrix of ligand-receptor pair names}

\item{type_a}{Cell type A}

\item{type_b}{Cell type B}

\item{dir_in}{Input directory, contains scRNAseq files}

\item{dir_out}{Output directory}
}
\value{
NIL
}
\description{
Given cell types and the names of ligand-receptor pairs, generate node prizes
from intracellular edges (gene relevance, mutual information), compute a
cross-talk score between cell types, and write out a configuration file to be
processed by the PCST algorithm.
}
