% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ligands_human}
\alias{ligands_human}
\title{Human ligand-receptor pairs}
\format{
A data frame with two variables: \code{ligand} and \code{receptor}
}
\usage{
ligands_human
}
\description{
Contians the names of human ligand-receptor pairs
}
\keyword{datasets}
