% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{match_lr_pairs}
\alias{match_lr_pairs}
\title{Find Ligand Receptor Pairs in Matrix}
\usage{
match_lr_pairs(mat, lrp)
}
\arguments{
\item{mat}{A numerical matrix}

\item{lrp}{Ligand-receptor pair data.frame; see \code{CytoTalk::lrp_human} for
an example}
}
\description{
Find Ligand Receptor Pairs in Matrix
}
