% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{normalize_sparse}
\alias{normalize_sparse}
\title{scRNAseq Count Normalization}
\usage{
normalize_sparse(mat, scale.factor = 10000)
}
\arguments{
\item{mat}{An integer matrix}

\item{scale.factor}{A single number constant by which to scale the
transformed data by}
}
\description{
scRNAseq Count Normalization
}
