% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcg_human}
\alias{pcg_human}
\title{Human protein-coding genes (PCG)}
\format{
A large character vector
}
\usage{
pcg_human
}
\description{
Contains the names of human protein-coding genes
}
\keyword{datasets}
