% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcg_mouse}
\alias{pcg_mouse}
\title{Mouse protein-coding genes (PCG)}
\format{
A large character vector
}
\usage{
pcg_mouse
}
\description{
Contains the names of mouse protein-coding genes
}
\keyword{datasets}
