% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.R
\name{read_matrix_with_meta}
\alias{read_matrix_with_meta}
\title{Read scRNAseq Data Matrix and Metadata}
\usage{
read_matrix_with_meta(fpath_mat, fpath_meta, auto_transform = TRUE)
}
\arguments{
\item{fpath_mat}{The path of a file containing a scRNAseq data matrix}

\item{fpath_meta}{The path of a file contianing column metadata (cell types)}

\item{auto_transform}{Should count data be transformed if detected?}
}
\value{
A named list containing a sparse data matrix and cell type metadata
}
\description{
Read scRNAseq Data Matrix and Metadata
}
\examples{
{
fpath_mat <- "~/Tan-Lab/scRNAseq-data-cpdb/sample_counts.txt"
fpath_meta <- "~/Tan-Lab/scRNAseq-data-cpdb/sample_meta.txt"
# lst_scrna <- CytoTalk::read_matrix_with_meta(fpath_mat, fpath_meta)
# result <- str(lst_scrna$cell_types)
}

}
