% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{subset_non_zero}
\alias{subset_non_zero}
\title{Subset Rows on Proportion Non-Zero}
\usage{
subset_non_zero(mat, cutoff)
}
\arguments{
\item{mat}{A numerical matrix}

\item{cutoff}{Threshold value in range [0, 1]; proportion of each row
required to be non-zero}
}
\description{
Subset Rows on Proportion Non-Zero
}
