% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{subset_rownames}
\alias{subset_rownames}
\title{Subset Rows on Rownames}
\usage{
subset_rownames(mat, labels)
}
\arguments{
\item{mat}{A numerical matrix}

\item{labels}{A subset of the matrix's rownames to filter by}
}
\description{
Subset Rows on Rownames
}
