% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{ungroup_meta}
\alias{ungroup_meta}
\title{Named List to Meta List}
\usage{
ungroup_meta(lst)
}
\arguments{
\item{lst}{A meta matrix list (outputted from \code{doc_fileio} methods)}
}
\description{
Named List to Meta List
}
