% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_gene_network.R
\name{write_integrated_net}
\alias{write_integrated_net}
\title{Write Out Integrated Network}
\usage{
write_integrated_net(dir_out)
}
\arguments{
\item{dir_out}{Output directory}
}
\value{
None
}
\description{
Uses node prize and edge cost data previously generated to write out a final
network configuration that can be processed by the PCST algorithm.
}
\examples{
\dontrun{
dir_out <- "my-output"
write_integrated_net(dir_out)
}

}
