% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_gene_network.R
\name{compute_crosstalk}
\alias{compute_crosstalk}
\title{Compute Cross-Talk Score}
\usage{
compute_crosstalk(type_a, type_b, dir_out)
}
\arguments{
\item{type_a}{Cell type A}

\item{type_b}{Cell type B}

\item{dir_out}{Output directory}
}
\value{
NIL
}
\description{
Compute cross-talk scores between cell type ligand-receptor pairs.
}
