% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mutual_information.R
\name{compute_mutual_information}
\alias{compute_mutual_information}
\title{Compute Mutual Information Matrix}
\usage{
compute_mutual_information(dir_out, cores = NULL)
}
\arguments{
\item{dir_out}{Output directory}

\item{cores}{Sets the number of cores}
}
\value{
None
}
\description{
Within each cell type, compute intracellular mutual information.
}
\examples{
{
dir_out <- "~/CytoTalk-output"
compute_mutual_information(dir_out, cores = 2)
}

}
