% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_indirect_edges.R
\name{filter_indirect_edges}
\alias{filter_indirect_edges}
\title{Filter Out Indirect Edges}
\usage{
filter_indirect_edges(dir_out)
}
\arguments{
\item{dir_out}{Output directory}
}
\value{
NIL
}
\description{
Using the ARACNE.M function, filter edges so that only direct edges remain.
}
