% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{ungroup_meta}
\alias{ungroup_meta}
\title{Named List to Meta List}
\usage{
ungroup_meta(lst)
}
\arguments{
\item{lst}{A meta matrix list (outputted from \code{doc_fileio} methods)}
}
\value{
A list of a matrix and a meta vector
}
\description{
Named List to Meta List
}
\examples{
{
lst_scrna <- CytoTalk::scrna_cyto
lst_group <- group_meta(lst_scrna)
result <- ungroup_meta(lst_group)
}

}
